watt issue on FIT

+++ src/sessionline.cpp
@@ -1,6 +1,6 @@
 #include "sessionline.h"
 
-SessionLine::SessionLine(double speed, int8_t inclination, double distance, uint8_t watt, int8_t resistance, uint8_t heart, double pace, uint8_t cadence, double calories, double elevationGain, QDateTime time)
+SessionLine::SessionLine(double speed, int8_t inclination, double distance, uint16_t watt, int8_t resistance, uint8_t heart, double pace, uint8_t cadence, double calories, double elevationGain, QDateTime time)
 {
     this->speed = speed;
     this->inclination = inclination;
